autonomous/learning/performance_optimizer.py
"""
PerformanceOptimizer - Auto-tune de parámetros (cache TTL, retry, thresholds).
"""
from __future__ import annotations

import logging
from pathlib import Path
from typing import Any

logger = logging.getLogger(__name__)


def _load_config() -> dict:
    cfg_path = Path(__file__).resolve().parent.parent.parent / "config" / "autonomous.yaml"
    if not cfg_path.exists():
        return {}
    try:
        import yaml
        with open(cfg_path, encoding="utf-8") as f:
            return yaml.safe_load(f) or {}
    except Exception:
        return {}


class PerformanceOptimizer:
    """Auto-tune de parámetros. approval_required controlado por governance (governed=True, growth=False)."""

    def __init__(self, config: dict | None = None):
        self._config = config or _load_config().get("learning", {})
        self._history: list[dict] = []
        self.approval_required: bool = self._config.get("approval_required", True)

    def optimize_parameter(self, param_name: str) -> Any:
        """Sugiere valor óptimo para un parámetro (por ahora devuelve None = no cambiar)."""
        return None

    def run_optimization_cycle(self) -> dict[str, Any]:
        """Ejecuta un ciclo de optimización sobre params conocidos."""
        return {"applied": [], "suggestions": []}

    def get_optimization_history(self) -> list[dict]:
        return list(self._history)

    def suggest_optimizations(self) -> list[str]:
        """Sugerencias sin aplicar."""
        return []
