requirements.txt
# Atlas Product Pack - core runtime
fastapi>=0.129.0
uvicorn>=0.41.0
websockets>=16.0
pydantic>=2.12.5
python-dotenv>=1.0.0
httpx>=0.28.1
python-multipart>=0.0.22

# ATLAS AUTONOMOUS (Health, Telemetry, Learning)
psutil>=7.2.2
scikit-learn>=1.3.0
py-cpuinfo>=9.0.0
PyYAML>=6.0

# Cognitive Architecture - Memory (embeddings + FAISS)
sentence-transformers>=3.0.0
faiss-cpu>=1.7.0

# Cognitive Architecture - Vision (depth, detection, segmentation)
timm>=0.9.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0  # YOLOv8

# Cognitive Architecture - Audio/Speech
openai-whisper>=20231117
sounddevice>=0.4.6
webrtcvad>=2.0.10

# Cognitive Architecture - Communication Bus
pyzmq>=25.0.0

# Cognitive Architecture - NLU/LLM
ollama>=0.1.0

# Memory Stack - ChromaDB + LangChain + LangFlow + Mem0 (from repo)
-e ./mem0_repo
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-chroma>=0.2.0
langflow>=1.7.0

# Compatibility fixes
pydantic-core>=2.41.5

# Cognitive Architecture - Graphs and Planning
networkx>=3.0

# Sandbox/Docker
docker>=6.1.0

# Observability (Prometheus)
prometheus_client>=0.19.0

# MediaPipe for pose estimation (optional)
# mediapipe>=0.10.0
